Made it more readable

+++ Firmware0_1/src/sensor/SensorReading.c
@@ -27,12 +27,16 @@ float sensor_getTemp(void){
 	 return( hih8120_getTemperature());
 }
 
+
 float sensor_getHum(void){
+
 	sensor_wakeUp();
+
 	_delay_ms(60);
+
 	hih8120_driverReturnCode_t measure_rc = hih8120_measure();
-	while (!(hih8120_isReady()))
-	{
+
+	while (!(hih8120_isReady())) {
 
 		_delay_ms(60);
 	}
@@ -41,11 +45,13 @@ float sensor_getHum(void){
 
 float sensor_getco2(void){
 	
-	  mh_z19_returnCode_t co2_rc = mh_z19_takeMeassuring();
-	  printf("MH_Z19 Measure status: %c\n", co2_rc);
-		 uint16_t ppm;
-	  mh_z19_getCo2Ppm(&ppm);
+	mh_z19_returnCode_t co2_rc = mh_z19_takeMeassuring();
+
+	printf("MH_Z19 Measure status: %c\n", co2_rc);
+
+	uint16_t ppm;
+	mh_z19_getCo2Ppm(&ppm);
 	  
-	  return ppm;
+	return (float)ppm;
 }
